requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
aiohttp
python-dateutil
pydantic-settings
